Lua errors now doesn't crash the program and fixed type conversion warnings

+++ include/ScriptHandlerInternal.h
@@ -32,7 +32,7 @@ namespace ScriptHandlerInternal
 	}
 	static int GetKeyNameFromCode(lua_State* ls) //keyName = GetKeyNameFromCode(keycode)
 	{
-		int keyCode = luaL_checknumber(ls, 1);
+		int keyCode = static_cast<int>(luaL_checknumber(ls, 1));
 		const char* name = SDL_GetKeyName(keyCode);
 		lua_pushstring(ls, name);
 		return 1;
@@ -55,9 +55,9 @@ namespace ScriptHandlerInternal
 
 	static int AddEntity(lua_State* ls) //id = AddEntity(type, textureFilename, radius)
 	{
-		int type = luaL_checknumber(ls, 1);
+		int type = static_cast<int>(luaL_checknumber(ls, 1));
 		std::string textureFilename(luaL_checkstring(ls, 2));
-		float radius = luaL_checknumber(ls, 3);
+		float radius = static_cast<float>(luaL_checknumber(ls, 3));
 		int id = entityHandler->Add(static_cast<EntityType>(type), textureFilename, radius, assetHandler);
 		lua_pushnumber(ls, id);
 		return 1;
@@ -66,7 +66,7 @@ namespace ScriptHandlerInternal
 	//Entity functions
 	static int GetEntityPosition(lua_State* ls) //x,y = GetEntityPosition(id)
 	{
-		int id = luaL_checknumber(ls, 1);
+		int id = static_cast<int>(luaL_checknumber(ls, 1));
 		Vec2f pos = entityHandler->GetEntity(id)->GetPosition();
 		lua_pushnumber(ls, pos.x);
 		lua_pushnumber(ls, pos.y);
@@ -74,7 +74,7 @@ namespace ScriptHandlerInternal
 	}
 	static int GetEntityVelocity(lua_State* ls) //xvel,yvel = GetEntityVelocity(id)
 	{
-		int id = luaL_checknumber(ls, 1);
+		int id = static_cast<int>(luaL_checknumber(ls, 1));
 		Vec2f vel = entityHandler->GetEntity(id)->GetVelocity();
 		lua_pushnumber(ls, vel.x);
 		lua_pushnumber(ls, vel.y);
@@ -82,7 +82,7 @@ namespace ScriptHandlerInternal
 	}
 	static int GetEntityAcceleration(lua_State* ls) //x,y = GetEntityAcceleration(id)
 	{
-		int id = luaL_checknumber(ls, 1);
+		int id = static_cast<int>(luaL_checknumber(ls, 1));
 		Vec2f acc = entityHandler->GetEntity(id)->GetAcceleration();
 		lua_pushnumber(ls, acc.x);
 		lua_pushnumber(ls, acc.y);
@@ -91,32 +91,32 @@ namespace ScriptHandlerInternal
 
 	static int SetEntityPosition(lua_State* ls) //SetEntityPosition(id, x, y)
 	{
-		int id = luaL_checknumber(ls, 1);
-		float x = luaL_checknumber(ls, 2);
-		float y = luaL_checknumber(ls, 3);
+		int id = static_cast<int>(luaL_checknumber(ls, 1));
+		float x = static_cast<float>(luaL_checknumber(ls, 2));
+		float y = static_cast<float>(luaL_checknumber(ls, 3));
 		entityHandler->GetEntity(id)->SetPosition(Vec2f(x, y));
 		return 0;
 	}
 	static int SetEntityVelocity(lua_State* ls) //SetEntityVelocity(id, xvel, yvel)
 	{
-		int id = luaL_checknumber(ls, 1);
-		float xvel = luaL_checknumber(ls, 2);
-		float yvel = luaL_checknumber(ls, 3);
+		int id = static_cast<int>(luaL_checknumber(ls, 1));
+		float xvel = static_cast<float>(luaL_checknumber(ls, 2));
+		float yvel = static_cast<float>(luaL_checknumber(ls, 3));
 		entityHandler->GetEntity(id)->SetVelocity(Vec2f(xvel, yvel));
 		return 0;
 	}
 	static int SetEntityAcceleration(lua_State* ls) //SetEntityAcceleration(id, xacc, yacc)
 	{
-		int id = luaL_checknumber(ls, 1);
-		float xacc = luaL_checknumber(ls, 2);
-		float yacc = luaL_checknumber(ls, 3);
+		int id = static_cast<int>(luaL_checknumber(ls, 1));
+		float xacc = static_cast<float>(luaL_checknumber(ls, 2));
+		float yacc = static_cast<float>(luaL_checknumber(ls, 3));
 		entityHandler->GetEntity(id)->SetAcceleration(Vec2f(xacc, yacc));
 		return 0;
 	}
 
 	static int GetEntityRadius(lua_State* ls) //r = GetEntityRadius(id)
 	{
-		int id = luaL_checknumber(ls, 1);
+		int id = static_cast<int>(luaL_checknumber(ls, 1));
 		float r = entityHandler->GetEntity(id)->GetRadius();
 		lua_pushnumber(ls, r);
 		return 1;
